refactor: create folder to make it clearer
