--- src/application/use-cases/place-order-output.ts
@@ -1,5 +1,0 @@
-export default class PlaceOrderOutput {
-
-    constructor (readonly code: string, readonly total: number) {
-    }
-}
